+++ src/models/train_mensa_multi_event_mixtures.py
@@ -73,13 +73,13 @@ if __name__ == "__main__":
                                               num_features, as_array=True)
     train_dict['X'] = torch.tensor(X_train, device=device, dtype=dtype)
     train_dict['E'] = torch.tensor(train_dict['E'], device=device, dtype=torch.int32)
-    train_dict['T'] = torch.tensor(train_dict['T'], device=device, dtype=torch.int32)
+    train_dict['T'] = torch.tensor(train_dict['T'], device=device, dtype=torch.float32)
     valid_dict['X'] = torch.tensor(X_valid, device=device, dtype=dtype)
     valid_dict['E'] = torch.tensor(valid_dict['E'], device=device, dtype=torch.int32)
-    valid_dict['T'] = torch.tensor(valid_dict['T'], device=device, dtype=torch.int32)
+    valid_dict['T'] = torch.tensor(valid_dict['T'], device=device, dtype=torch.float32)
     test_dict['X'] = torch.tensor(X_test, device=device, dtype=dtype)
     test_dict['E'] = torch.tensor(test_dict['E'], device=device, dtype=torch.int32)
-    test_dict['T'] = torch.tensor(test_dict['T'], device=device, dtype=torch.int32)
+    test_dict['T'] = torch.tensor(test_dict['T'], device=device, dtype=torch.float32)
     
     n_samples = train_dict['X'].shape[0]
     n_features = train_dict['X'].shape[1]
