+++ src/tuning/tune_mtlr_model.py
@@ -74,13 +74,13 @@ def train_model():
                                               num_features, as_array=True)
     train_dict['X'] = torch.tensor(X_train, device=device, dtype=dtype)
     train_dict['E'] = get_first_event(torch.tensor(train_dict['E'], device=device, dtype=torch.int32))
-    train_dict['T'] = get_first_event(torch.tensor(train_dict['T'], device=device, dtype=torch.int32))
+    train_dict['T'] = get_first_event(torch.tensor(train_dict['T'], device=device, dtype=torch.float32))
     valid_dict['X'] = torch.tensor(X_valid, device=device, dtype=dtype)
     valid_dict['E'] = get_first_event(torch.tensor(valid_dict['E'], device=device, dtype=torch.int32))
-    valid_dict['T'] = get_first_event(torch.tensor(valid_dict['T'], device=device, dtype=torch.int32))
+    valid_dict['T'] = get_first_event(torch.tensor(valid_dict['T'], device=device, dtype=torch.float32))
     test_dict['X'] = torch.tensor(X_test, device=device, dtype=dtype)
     test_dict['E'] = get_first_event(torch.tensor(test_dict['E'], device=device, dtype=torch.int32))
-    test_dict['T'] = get_first_event(torch.tensor(test_dict['T'], device=device, dtype=torch.int32))
+    test_dict['T'] = get_first_event(torch.tensor(test_dict['T'], device=device, dtype=torch.float32))
 
     # Make time bins
     time_bins = make_time_bins(train_dict['T'], event=None, dtype=dtype).to(device)
